requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.0
azure-ai-vision==0.15.1b1
azure-identity==1.15.0
PyMuPDF==1.23.8
Pillow==10.2.0 